stod-backend/requirements.txt
asgiref==3.2.10
cffi==1.14.5
cryptography==3.2.1
Django==3.1.2
django-cors-headers==3.5.0
django-crispy-forms==1.11.1
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.12.1
drf-pretty-exception-handler==0.1.1
gunicorn==20.0.4
psycopg2==2.8.6
pycparser==2.20
pytz==2020.1
signals==0.0.2
six==1.15.0
sqlparse==0.4.1
